pyproject.toml
[project]
name = "fastapiproj"
version = "0.1.0"
description = ""
authors = [
    {name = "tomcat",email = "tomcat@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0" 
dependencies = [
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "tzdata (>=2025.1,<2026.0)"
]

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.4"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
taskipy = "^1.14.1"

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev fastapiproj/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=fastapiproj -vv'
post_test = 'coverage html'
alembicrev = 'alembic revision --autogenerate -m "create users table'
alembichead = 'alembic upgrade head'

[tool.pytest.ini_options]
pythonpath = "."
addopts = "-p no:warnings"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


